# docs/openapi/v1.yaml
openapi: 3.1.0
info:
  title: Agenda Barbearia API
  version: 1.1.0
  summary: Integrações de agenda (n8n, WhatsApp e app do cliente)
  description: |
    API REST para **conectar o WhatsApp, o n8n ou o app do cliente** à agenda da
    barbearia — as mesmas regras do site `/agenda`.

    **O que dá para fazer**
    - Ver dados da **loja** (`/shop`)
    - Ver **serviços** e **profissionais**
    - Consultar **horários livres**, **criar**, **listar**, **remarcar** e **cancelar**
    - Buscar **cliente** pelo WhatsApp
    - No **app**: ver e **editar o próprio cadastro** (`GET` / `PATCH /customers/me`),
      registrar push (`/customers/me/push-token`), ver notificações (`/customers/me/notifications`)
      e **sair de todos os aparelhos** (`DELETE /customers/me/sessions`)
    - Disparar e confirmar **lembretes** (~1h antes)
    - No **n8n**: ativar/pausar a **IA** por conversa (`/ai-status`)

    Comandas, caixa e comissões ficam **só no painel** — não entram nesta API.

    ## Repetir sem duplicar (Idempotency-Key)

    Em `POST /appointments`, `PATCH /appointments/{id}` e `DELETE /appointments/{id}`,
    envie um header `Idempotency-Key` (qualquer string única por tentativa,
    ex.: um UUID). Se a chamada for repetida com a mesma chave — por exemplo,
    um retry automático depois de um timeout — o servidor devolve a mesma
    resposta da primeira vez em vez de repetir a ação.

    ## Autenticação

    ### Chave de API (n8n / integrações)

...

    ```
...
    ```

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - BearerAuth: []
paths:
...
  /services:
    get:
      tags: [Serviços]
      operationId: listServices
      summary: Listar serviços ativos
      description: |
        Lista os serviços que a barbearia oferece hoje: nome, duração, foto,
        preços por dia da semana e quem realiza cada um.

        Opcional:
        - `professionalId` — só o que aquele barbeiro faz
        - `date` — preenche o preço daquele dia (`priceCentsForDate`)

        **Acesso:** pública. Com chave, scope `catalog:read`.
      security: []
      parameters:
        - name: professionalId
          in: query
          description: Filtra pelos serviços que esse barbeiro realiza
          schema:
            type: string
            format: uuid
        - name: date
          in: query
          description: Dia AAAA-MM-DD para calcular o preço da data
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Lista de serviços
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  timezone:
                    type: string
                    example: America/Sao_Paulo
                  dayLabels:
                    type: array
                    items:
                      type: string
                    example: [Dom, Seg, Ter, Qua, Qui, Sex, Sab]
                  date:
                    type: string
                    nullable: true
                    format: date
                  weekday:
                    type: integer
                    nullable: true
                  services:
                    type: array
                    items:
                      $ref: "#/components/schemas/Service"
              example:
                ok: true
                timezone: America/Sao_Paulo
                dayLabels: [Dom, Seg, Ter, Qua, Qui, Sex, Sab]
                date: null
                weekday: null
                services:
                  - id: 9f135f0a-6894-4062-91df-06233c18d10e
                    name: Barbaterapia
                    description: ""
                    durationMinutes: 30
                    priceCents: 6000
                    priceCentsForDate: null
                    priceFrom: false
                    prices:
                      - priceCents: 6000
                        weekdays: [1, 2, 3]
                        days: [Seg, Ter, Qua]
                      - priceCents: 6500
                        weekdays: [4, 5, 6]
                        days: [Qui, Sex, Sab]
                    photoUrl: null
                    professionalIds:
                      - 054a545a-75c8-4807-b72d-5c460bb3539f
                    professionals:
                      - id: 054a545a-75c8-4807-b72d-5c460bb3539f
                        nickname: Junior
                        photoUrl: null
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /professionals:
    get:
      tags: [Profissionais]
      operationId: listProfessionals
      summary: Listar profissionais ativos
      description: |
        Lista os barbeiros ativos para escolha no agendamento: apelido, foto
        e os IDs dos serviços que cada um realiza (`serviceIds`).

        Detalhes de preço e duração dos serviços ficam em `GET /services`.
        Opcional: `serviceId` para listar só quem faz aquele serviço.

        **Acesso:** pública. Com chave, scope `catalog:read`.
      security: []
      parameters:
        - name: serviceId
          in: query
          description: Filtra pelos barbeiros que realizam este serviço
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Lista de profissionais
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  timezone:
                    type: string
                  professionals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Professional"
              example:
                ok: true
                timezone: America/Sao_Paulo
                professionals:
                  - id: 054a545a-75c8-4807-b72d-5c460bb3539f
                    nickname: Junior
                    photoUrl: null
                    serviceIds:
                      - 3a62b091-6916-4741-a3d4-754a33b2cb31
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /appointments:
    post:
      tags: [Agendamentos]
      operationId: createAppointment
      summary: Criar agendamento
      description: |
        Marca um novo horário para o cliente (nome, WhatsApp, serviços, data e hora).

        Antes de chamar, use `GET /appointments/availability` para garantir
        que o horário ainda está livre. O servidor valida de novo ao gravar.

        Informe o barbeiro (`professionalId`) **ou** `anyProfessional: true`
        para o sistema escolher alguém disponível.

        **Acesso:** privada. Scope `appointments:create`
        (chave de API, cookie OTP ou `Bearer` com `accessToken` do OTP).

        Recomendado enviar `Idempotency-Key` (ver parâmetro) pra evitar
        marcar o mesmo horário duas vezes num retry.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/idempotencyKeyHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentBody"
            example:
              anyProfessional: true
              date: "2026-07-23"
              startTime: "10:00"
              serviceIds:
                - 3a62b091-6916-4741-a3d4-754a33b2cb31
              firstName: João
              lastName: Silva
              whatsapp: "11999999999"
      responses:
        "200":
          description: Agendamento criado
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  appointmentId:
                    type: string
                    format: uuid
                  professionalId:
                    type: string
                    format: uuid
                  professionalNickname:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    get:
      tags: [Agendamentos]
      operationId: listAppointments
      summary: Listar agendamentos do cliente
      description: |
        Lista os agendamentos de um cliente pelo WhatsApp.

        - `mode=upcoming` (padrão): só futuros ativos — usado pelo site "Meus horários"
        - `mode=history`: passados, cancelados e concluídos
        - `mode=all`: todos os status (até 50)

        **Acesso:** privada. Scope `appointments:read`
        (chave de API, dono, cookie OTP ou `Bearer` com `accessToken`).
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/whatsappQuery"
        - name: mode
          in: query
          description: upcoming | history | all
          schema:
            type: string
            enum: [upcoming, history, all]
            default: upcoming
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  mode:
                    type: string
                    enum: [upcoming, history, all]
                  appointments:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicAppointment"
              example:
                ok: true
                mode: upcoming
                appointments:
                  - id: 11111111-1111-1111-1111-111111111111
                    professionalId: 054a545a-75c8-4807-b72d-5c460bb3539f
                    professionalName: Junior
                    professionalPhotoUrl: null
                    date: "2026-07-23"
                    startTime: "10:00"
                    status: scheduled
                    serviceIds:
                      - 3a62b091-6916-4741-a3d4-754a33b2cb31
                    serviceNames: [Corte]
                    totalMinutes: 30
                    totalPriceCents: 6000
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: dbc_live_… ou accessToken do OTP
      description: |
        Dois tipos de Bearer:
        1. **Chave de API** (`dbc_live_<keyId>_<secret>`) — n8n / integrações
        2. **Sessão do cliente** (`accessToken` de `POST /api/agenda/otp/verify`) — app mobile

        Exemplo: `Authorization: Bearer <token>`

        `/customers/me` aceita **somente** o token de sessão do cliente (não chave de API).
  parameters:
    whatsappQuery:
      name: whatsapp
      in: query
      required: true
      schema:
        type: string
      example: "11999999999"
      description: WhatsApp com DDD (com ou sem 55)
...
    idempotencyKeyHeader:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Chave opcional escolhida por quem chama (ex.: um UUID gerado antes
        de enviar). Se a mesma chamada for repetida com a mesma chave — por
        exemplo, um retry por timeout de rede do app ou do n8n — o servidor
        devolve a resposta salva da primeira vez, em vez de repetir a ação
        (evita agendamento duplicado). Reusar a chave com dados diferentes
        dá **409**.
      schema:
        type: string
        maxLength: 200
  schemas:
    Error:
      type: object
      description: |
        Formato de erro padrão de toda a API: `ok` sempre `false` e `error`
        com uma mensagem em português pronta pra mostrar pro usuário.
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: string
    Professional:
      type: object
      description: |
        Dados públicos do profissional. Contato interno (e-mail, WhatsApp do barbeiro),
        nome/sobrenome e comissão não entram nesta rota pública.
      properties:
        id:
          type: string
          format: uuid
        nickname:
          type: string
        photoUrl:
          type: string
          nullable: true
        serviceIds:
          type: array
          description: IDs dos serviços que realiza (detalhes em GET /services)
          items:
            type: string
            format: uuid
    Service:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        durationMinutes:
          type: integer
        priceCents:
          type: integer
          description: Menor preço da semana (referência)
        priceCentsForDate:
          type: integer
          nullable: true
          description: Preço do dia em ?date= (null sem date)
        priceFrom:
          type: boolean
        prices:
          type: array
          description: Preços agrupados por valor
          items:
            type: object
            properties:
              priceCents:
                type: integer
              weekdays:
                type: array
                items:
                  type: integer
                description: "0=Dom … 6=Sáb"
              days:
                type: array
                items:
                  type: string
                example: [Seg, Ter, Qua]
        photoUrl:
          type: string
          nullable: true
        professionalIds:
          type: array
          description: IDs dos profissionais que realizam este serviço
          items:
            type: string
            format: uuid
        professionals:
          type: array
          description: Resumo público de quem realiza (sem serviceIds aninhados)
          items:
            type: object
            properties:
              id:
                type: string
                format: uuid
              nickname:
                type: string
              photoUrl:
                type: string
                nullable: true
...
    PublicAppointment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        professionalId:
          type: string
          format: uuid
        professionalName:
          type: string
        professionalPhotoUrl:
          type: string
          nullable: true
        date:
          type: string
          format: date
        startTime:
          type: string
        status:
          type: string
          enum: [scheduled, confirmed, cancelled, done]
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
        serviceNames:
          type: array
          items:
            type: string
        totalMinutes:
          type: integer
        totalPriceCents:
          type: integer
    CreateAppointmentBody:
      type: object
      required: [date, startTime, serviceIds, firstName, lastName, whatsapp]
      properties:
        professionalId:
          type: string
          format: uuid
        anyProfessional:
          type: boolean
        date:
          type: string
          format: date
        startTime:
          type: string
          example: "10:00"
        serviceIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        firstName:
          type: string
        lastName:
          type: string
        whatsapp:
          type: string
...
  responses:
    BadRequest:
      description: Parâmetros inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Não autorizado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            error: Não autorizado.
    Forbidden:
      description: Sem permissão
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            error: Sem permissão.
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: Conflito de regra de negócio
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit excedido
      headers:
        Retry-After:
          schema:
            type: integer
          description: Segundos até tentar de novo
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Muitas tentativas em pouco tempo. Aguarde alguns minutos e tente de novo.
...